date/time improvements and fixes
The regular format procedure always returns the formatted time in the UTC time zone. In the new procedure, you will be able to specify your desired time zone as an additional argument. The time zone can be in any format accepted by the java.util.TimeZone.getTimeZone(String) method.

Tests for formatTimeZone procedure

Add a description of the new formatTimeZone procedure

Fix the date test cases

It would be semantically incorrect to test with a format that doesn't include month and date. Previous Travis build failed because of this.

Fixed Date/timezone bahavior