Strict mode for apoc.meta.subGraph
Enabling strict mode restricts the meta graph to only include the specified labels, and not any other. Even if a given relationship type may link to other labels, they will not be included in the meta graph.

Interestingly, not specifying any relationship type will show all relationship types applicable to the specified set of labels. Not specifying any labels will show all labels, but only the specified relationship types.

Default value is false, to maintain the existing behavior.

In strict mode, empty label list should return only the labels connected to given relationship types. This changes the previous (and lenient) behavior where it returns all labels.

In strict mode, relationships are only added if the labels connected by them are allowed. Since labels are added by sampling, if a relationship is sampled before its corresponding labels are sampled, there is a chance that the relationship may never be added to the meta graph.

The fix is to always add the labels of a relationship first (if allowed) before adding the relationship itself.

Since adding labels incremented the count, the previous change introduced another subtle bug. The counts were all wrong! This fix resolves the problem by making sure the addition of labels before relationships will not increment the counts of such labels.