Fix the partitioning logic in mapParallel partitionSubList
The original partitioning logic is wrong and causes AssertionError when doing testMapParallel. The fix is to get ceiling of the quotient so that the partitions will always cover the total number of items in the list. In addition, the subList call had to include another Math.min call to avoid IndexOutOfBoundsException.

Example:
Total: 1000
Partitions: 400

Incorrect batch size logic: 1000/400 = (int)2
With each batch having 2 list elements, the subList can only reach to (2*400=) 800 elements.

Correct batch size logic: 1000/400 = 2.5; ceiling of 2.5 = 3
With each batch having 3 list elements, the subList can reach up to (3*400=) 1200 elements, enough to cover the total of 1000 elements in the list.