Enhance help to search whole procedure name, and be case insensitive
The original search was too restrictive as it only checked the beginning and end of the string, and moreover it was case sensitive. This makes the search more lenient by allowing one to search anywhere within the procedure name, and also ignores the case of the search terms.

This gives more relevant results, such as searching for all procedures related to "meta" or getting information about "apoc.meta.subGraph" when searching for "graph".