Add ability to search the description text by appending + to search term
When a + sign is found at the end of the search string, the help system also searches the descriptive text of the procedures. This is useful to return any procedures whose descriptions mention the search term.

The change also trims the search string so any spaces at the beginning and end of the string are ignored.