Fix the Help System (#48)
* Use the class's own ClassLoader to find the apoc jar file

When using thread's context class loader, it is not guaranteed to find the apoc jar. When using the class's own class loader, it is guaranteed, since this class comes from apoc jar itself.

This ensures the help system always works.

* Enhance help to search whole procedure name, and be case insensitive

The original search was too restrictive as it only checked the beginning and end of the string, and moreover it was case sensitive. This makes the search more lenient by allowing one to search anywhere within the procedure name, and also ignores the case of the search terms.

This gives more relevant results, such as searching for all procedures related to "meta" or getting information about "apoc.meta.subGraph" when searching for "graph".

* Add ability to search the description text by appending + to search term

When a + sign is found at the end of the search string, the help system also searches the descriptive text of the procedures. This is useful to return any procedures whose descriptions mention the search term.

The change also trims the search string so any spaces at the beginning and end of the string are ignored.

* Better documentation for the apoc.help procedure

Includes instructions on searching the description text