Added new Gradle task to run asciidoctor and build the index.html file
The additional copy task is necessary as asciidoctor cannot directly output to the root of the project in a Windows system due to file locking. Having it output to a build folder then copying from there ensures that the task will run in any system.

***** Removed the shell script, as it's no longer needed. *****

Note that the index.html is 100% generated by asciidoctor