Set the default locale for tests
Some tests (NumberTest) format number using the default locale, and assert
that results are formatted using the C or en_US locale (point as the decimal
separator, comma as the grouping separator, etc.). The tests would fail
when run by a user with a locale with different settings (fr_FR for
example).