Use their max ids to warm up all the nodes/rels (#131)
Because the nodes (or relationships) might have been deleted and their ids
not reused, simply using the count is not enough to traverse the complete
id space: instead the maximum id needs to be fetched from the store.

This commit mostly reverts to the original pull request, without the
hard-coded values.