Added apoc.data.url.* and apoc.data.email.* parse functions
* ExtractEmail functions: apoc.data.email.(domain, user, personal)
* first commit of ExtractURL and associated tests
* deprecated Extract class in favor of apoc.data.url.* and apoc.data.email.*

This PR is intended to address #664

Additionally it adds new functions mapping to standard Java API provided functions which are easy and useful to expose.

All of the original test cases that were used as part of the original Extract class have been preserved and new test cases have been added to cover the new functions added.