Avoid using recursion for re-submitting rejected tasks in the Pools. (#1193)
The RejectedExecutionHandler used by the APOC Pools for all sorts of asynchronisity and parallelism, was attempting to block the submitting thread until the task was accepted by the executor.
This was done by letting the rejected execution handler submit the rejected task to the executor again. If the task was once again rejected, the handler would be called again, and so on.
The troubling part of this is that this happened recursively. This could produce some enourmous stack depths for many threads.
Some threads would get StackOverflowErrors when trying to submit tasks, which could break invariants and compromise the integrity of the system.

The solution is to make use of the blocking APIs on the internal task queue of the executor, to manually enqueue tasks in a blocking way.
This is done using the blocking APIs that take a timeout, such that we can still periodically check if the pool has been shut down.