Fixes some issues with R4L property observations. Previously properties were coming back which weren't actually assigned due to the seperation of relType and source/target nodes. This fix combines all three parameters in to one identifier per relationship.
Note: Observations are counted by label (not node) so if a node has multiple labels, the number of observations will be higher. This should not affect the accuracy of data, but may have a slight performance affect that could become a bigger issue with large graphs that routinely use multiple labels per node. This issue is scheduled to be resolved in a future version.