Only register with kernel if there are triggers
The kernel event handling creates a big expensive transaction snapshot
for every transaction if there are registered event handlers.

To avoid paying that overhead, unregister from the kernel if there
are no triggers.

This works as a complement to the config option, so you can deploy
large groups of databases with the config enabled but only pay the
overhead on those databases that use trigger.