Refactor the config file loading and add tests
Back off from the LoadJdbc literal adaption and eliminate the
credentials map. Just make username (instead of user) and password
keys in the connectionMap
Also collapse the LoadLdapConfig constructors to streamline the object
construction. Hopefully, this will make a backwards-compatibility with
the old plugin configuration style easier to do
Also, better handle cases of bad config values for pageSize. If
parseInt is unble to make sense of it, fallback to the default value
of 100