Handle Optional.empty() function descriptions
This handles loading a stored APOC custom state where
a function or procedure had an Optional.empty() description.

Jackson serialized those to {"present::false}, which then broke
Apocs assumptions about the schema here.

This makes it so we handle any in-the-wild incarnations of this,
and modifies our write path to always ensure the value is a string