Support GCP environmental service account auth
Previously, we forced using service account private keys set using the
GOOGLE_APPLICATION_CREDENTIALS environment variable pointing to a
generated json private key. This isn't required in native GCP
environments as the official Google Cloud clients will identify if
they're running in a GCP service (e.g. GAE, CloudRun, GCE, etc.) and
provide automatic authentication as a service account.

This change adds a new "authenticationType" and renames the old type:

- SERVICE becomes PRIVATE_KEY to explicitly describe that it's expecting
  to use a private key file.
- GCP_ENVIRONMENT is added to be used when running in GCP.

The logic is that for PRIVATE_KEY usage we still check the environment
for a set GOOGLE_APPLICATION_CREDENTIALS value. A runtime error is
thrown if it's not set.

For GCP_ENVIRONMENT, the environment check is ignored as the Google
client library will take care of validation.