fix apoc.periodic.iterate not obeying concurrency
The procedure uses the default pooled executor when running with
parallel=true, but the pool is #cpus * 2...so the concurrency ends up
being 2x the cpu cores and not what the user sets.

This changes the behavior to use a set cap via an AtomicInteger and
keeps the same Future creation/submission logic but adds the decrement
of the counter to the future's callback. If parallel == false, the
gating is shortcircuited at the expense of some small overhead for
manipulating the AtomicInteger counter (for now?).