Summarize graph updates for apoc.periodic.iterate
Initial fix for #1815 adding in support for providing aggregate update
statistics for any changes to the database from the
apoc.periodic.iterate consuming statement.

For now, the following data is provided:

- nodesCreated
- nodesDeleted
- relationshipsCreated
- relationshipsDeleted
- propertiesSet
- labelsAdded
- labelsRemoved

Consequently, this also adds the capability to the deprecated
apoc.periodic.rock_n_roll as it now uses shared logic.

Given the challenge of doing schema operations in the consuming
statement, index and constraint updates are not provided. Support might
be required for using apoc.schema.assert, but it will in turn need to be
updated to properly report update statistics.