Provide update statistics for apoc.periodic.iterate (issue #1815) (#1818)
* fix apoc.periodic.iterate not obeying concurrency

The procedure uses the default pooled executor when running with
parallel=true, but the pool is #cpus * 2...so the concurrency ends up
being 2x the cpu cores and not what the user sets.

This changes the behavior to use a set cap via an AtomicInteger and
keeps the same Future creation/submission logic but adds the decrement
of the counter to the future's callback. If parallel == false, the
gating is shortcircuited at the expense of some small overhead for
manipulating the AtomicInteger counter (for now?).

* Fix #1805: prevent execution with invalid params

To prevent executing with batch sizes of 0 or concurrency settings that
result in no consuming threads, validate that batchSize and concurrency
are set to at least 1.

* Use native line separators so tests work on Win10

* Centralize multithreaded batch processing logic

Both APOC Core's apoc.periodic.Periodic and Full's
apoc.periodic.PeriodicExtended had implementations of logic for batch
processing with thread pools. Pull it out and centralize it in
PeriodicUtils for now.

The method signature ends up growing quite a bit, but it's now in 1
place and apoc.periodic.rock_n_roll now had the same logic as
apoc.periodic.iterate.

* cleanup test class, fix error comparison for win10

* Summarize graph updates for apoc.periodic.iterate

Initial fix for #1815 adding in support for providing aggregate update
statistics for any changes to the database from the
apoc.periodic.iterate consuming statement.

For now, the following data is provided:

- nodesCreated
- nodesDeleted
- relationshipsCreated
- relationshipsDeleted
- propertiesSet
- labelsAdded
- labelsRemoved

Consequently, this also adds the capability to the deprecated
apoc.periodic.rock_n_roll as it now uses shared logic.

Given the challenge of doing schema operations in the consuming
statement, index and constraint updates are not provided. Support might
be required for using apoc.schema.assert, but it will in turn need to be
updated to properly report update statistics.

* update docs, adding updateStatistics yield value