Introduce Data Virtualization procedures
The APOC library now supports the definition of a catalog of virtual resources. 

A virtual resource is an external data source that neo4j can use to query and retrieve data on demand presenting it as virtual nodes enriching the data stored in the graph.

Virtual resources are currently limited to CSV and JDBC sources.