Remove usage of CoordinateReferenceSystem
`CoordinateReferenceSystem` is not listed as `@PublicAPI`. This usage was found in a failing build.
A pending internal PR is changing some constant names to fit with internal naming conventions. Rather than fix it to use the new (still internal) name; this change explicitly checks for the string, minimising direct use and assumptions, of internals.