Normalise Local and CI test environments
The test configurations would assign different resources depending on whether the tests were running locally or in the CI. This is usually considered bad practice as tests should run in a predictable manner and within environments as close as possible to avoid flakiness.
The test configurations have also been parallelised in order to attempt faster runs.