Enable tests to run in CI
Many tests were being ignored in the CI because they include assumeFalse(isRunningInCI()) statements which stop them from running. This means that they are not catching errors and regressions which can lead to a lower quality product. We would like these tests to run in the CI going forward even if it takes more time.