Run full tests in ci (#2819)
* Normalise Local and CI test environments

The test configurations would assign different resources depending on whether the tests were running locally or in the CI. This is usually considered bad practice as tests should run in a predictable manner and within environments as close as possible to avoid flakiness.
The test configurations have also been parallelised in order to attempt faster runs.

* Enable tests to run in CI

Many tests were being ignored in the CI because they include assumeFalse(isRunningInCI()) statements which stop them from running. This means that they are not catching errors and regressions which can lead to a lower quality product. We would like these tests to run in the CI going forward even if it takes more time.

* Cleanup test resources

The tests in APOC full would oftentimes not clean up resources after they had ran. This made it difficult to run them all in the CI because the CI would run out of resources (memory, threads, disk space).

* Ignore broken tests

As part of this change, we want to reestablish making APOC tests run in the CI but unfortunately we have discovered that some tests are broken. I am ignoring them for now and will create separate tickets to remove @Ignore annotations in the future since this is in itself quite a large change.