Avoid reusing the AmbientCursor in node.getLabels()
We were calling node.getLabels() severals which is unecessary work, and is also might have correctness implications because it is an AmbientCursor. During the bug, https://trello.com/c/zLF4jj16/194-s4sony-europe-network-map-attack-surface-underlyingstorageexception-access-to-record-node-went-out-of-bounds-of-the-page, we had a conversation with Sergey were he mentions that AmbientCursor reuse should be avoided.

(cherry picked from commit c8a96f3e298d34951882fa2375ce77926258d662)